add reg User

+++ DmitryGoncharov/Blog/src/main/java/by/epam/blog/model/User.java
@@ -1,16 +1,12 @@
 package by.epam.blog.model;
 
 import java.io.Serializable;
-import java.util.Collection;
-import java.util.LinkedHashSet;
 
-import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
 /**
@@ -25,7 +21,7 @@ public class User implements Serializable{
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int id;
+	private long id;
 	@Column(name = "USER_LOGIN")
 	private String login;
 	@Column(name = "USER_PASS")
@@ -58,7 +54,7 @@ public class User implements Serializable{
 	public void setName(String name) {
 		this.name = name;
 	}
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 	public void setId(int id) {
